+++ src/PermissionRegistrar.php
@@ -2,11 +2,11 @@
 
 namespace Spatie\Permission;
 
-use Illuminate\Support\Collection;
 use Illuminate\Contracts\Auth\Access\Gate;
-use Illuminate\Contracts\Cache\Repository;
-use Spatie\Permission\Contracts\Permission;
 use Illuminate\Contracts\Auth\Authenticatable;
+use Illuminate\Contracts\Cache\Repository;
+use Illuminate\Support\Collection;
+use Spatie\Permission\Contracts\Permission;
 use Spatie\Permission\Exceptions\PermissionDoesNotExist;
 
 class PermissionRegistrar
@@ -16,6 +16,9 @@ class PermissionRegistrar
 
     /** @var \Illuminate\Contracts\Cache\Repository */
     protected $cache;
+
+    /** @var \Illuminate\Contracts\Logging\Log */
+    protected $logger;
 
     /** @var string */
     protected $cacheKey = 'spatie.permission.cache';
@@ -28,16 +31,17 @@ class PermissionRegistrar
 
     public function registerPermissions(): bool
     {
-        $this->gate->before(function (Authenticatable $user, string $ability) {
+        $this->gate->before(function (Authenticatable $user, string $ability, $restrictable = null) {
             try {
                 if (method_exists($user, 'hasPermissionTo')) {
-                    return $user->hasPermissionTo($ability) ?: null;
+                    return $user->hasPermissionTo($ability, $restrictable) ?: null;
                 }
             } catch (PermissionDoesNotExist $e) {
             }
         });
 
         return true;
+
     }
 
     public function forgetCachedPermissions()
@@ -45,6 +49,8 @@ class PermissionRegistrar
         $this->cache->forget($this->cacheKey);
     }
 
+    // The getPermissions() cache implementation is used via the Permission's findByName() method, which is called by
+    // the hasPermissionTo() method (and the similar ones) of HasRoles Trait
     public function getPermissions(): Collection
     {
         return $this->cache->remember($this->cacheKey, config('permission.cache_expiration_time'), function () {